refac(utils): reorganized dex utils
also updated imports in some files
