+++ src/pages/Calcdex/PlayerCalc.tsx
@@ -8,8 +8,8 @@ import { eacute } from '@showdex/consts/core';
 import { useUserLadderQuery } from '@showdex/redux/services';
 import { useColorScheme } from '@showdex/redux/store';
 import { findPlayerTitle, formatId, openUserPopup } from '@showdex/utils/app';
-import { hasNickname } from '@showdex/utils/battle';
 import { env, getResourceUrl } from '@showdex/utils/core';
+import { hasNickname } from '@showdex/utils/dex';
 import { capitalize } from '@showdex/utils/humanize';
 import type { DropdownOption } from '@showdex/components/form';
 import type { ElementSizeLabel } from '@showdex/utils/hooks';
