+++ src/pages/Calcdex/PokeItemOptionTooltip.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import cx from 'classnames';
-import { formatDexDescription, getDexForFormat } from '@showdex/utils/battle';
+import { formatDexDescription, getDexForFormat } from '@showdex/utils/dex';
 import type { ItemName } from '@smogon/calc/dist/data/interface';
 import type { SelectOptionTooltipProps } from '@showdex/components/form';
 import styles from './PokeInfo.module.scss';
