+++ src/utils/battle/index.ts
@@ -8,22 +8,10 @@ export * from './detectPlayerKey';
 export * from './detectPokemonIdent';
 export * from './detectSpeciesForme';
 export * from './detectToggledAbility';
-export * from './formatDexDescription';
-export * from './getDexForFormat';
-export * from './getGenDexForFormat';
 export * from './getGenlessFormat';
-export * from './getMaxMove';
-export * from './getZMove';
-export * from './getNaturesDex';
-export * from './getPokemonLearnset';
-export * from './getTypesDex';
-export * from './getWeatherConditions';
 export * from './guessTableFormatKey';
-export * from './hasMegaForme';
-export * from './hasNickname';
 export * from './legalLockedFormat';
 export * from './mergeRevealedMoves';
-export * from './notFullyEvolved';
 export * from './ruinAbilitiesActive';
 export * from './sanitizeField';
 export * from './sanitizeMoveTrack';
@@ -32,7 +20,6 @@ export * from './sanitizePokemon';
 export * from './sanitizeVolatiles';
 export * from './syncField';
 export * from './syncPokemon';
-export * from './toggleableAbility';
 export * from './toggleRuinAbilities';
 export * from './usedDynamax';
 export * from './usedTerastallization';
