+++ src/utils/battle/mergeRevealedMoves.ts
@@ -1,8 +1,8 @@
 import { PokemonPivotMoves } from '@showdex/consts/pokemon';
+import { getDexForFormat } from '@showdex/utils/dex';
 import { flattenAlts } from '@showdex/utils/presets';
 import type { MoveName } from '@smogon/calc/dist/data/interface';
 import type { CalcdexPokemon } from '@showdex/redux/store';
-import { getDexForFormat } from './getDexForFormat';
 
 /**
  * Intelligently uses hardcoded intelligence to merge the `pokemon`'s currently set `moves`
