+++ src/utils/battle/sanitizePokemon.ts
@@ -1,6 +1,7 @@
 import { PokemonNatures } from '@showdex/consts/pokemon';
 import { calcPokemonCalcdexId } from '@showdex/utils/calc';
 import { env } from '@showdex/utils/core';
+import { getDexForFormat, toggleableAbility } from '@showdex/utils/dex';
 import { flattenAlts } from '@showdex/utils/presets';
 import type { GenerationNum } from '@smogon/calc';
 import type { AbilityName, ItemName, MoveName } from '@smogon/calc/dist/data/interface';
@@ -11,10 +12,8 @@ import { detectPlayerKeyFromPokemon } from './detectPlayerKey';
 import { detectPokemonIdent } from './detectPokemonIdent';
 import { detectSpeciesForme } from './detectSpeciesForme';
 import { detectToggledAbility } from './detectToggledAbility';
-import { getDexForFormat } from './getDexForFormat';
 import { sanitizeMoveTrack } from './sanitizeMoveTrack';
 import { sanitizeVolatiles } from './sanitizeVolatiles';
-import { toggleableAbility } from './toggleableAbility';
 
 /**
  * Essentially converts a `Showdown.Pokemon` into our custom `CalcdexPokemon`.
