+++ src/utils/battle/syncPokemon.ts
@@ -8,6 +8,7 @@ import {
 } from '@showdex/utils/calc';
 import { env } from '@showdex/utils/core';
 // import { logger } from '@showdex/utils/debug';
+import { getDexForFormat } from '@showdex/utils/dex';
 import { capitalize } from '@showdex/utils/humanize';
 import { flattenAlts, guessTeambuilderPreset } from '@showdex/utils/presets';
 import type { GenerationNum } from '@smogon/calc';
@@ -15,7 +16,6 @@ import type { AbilityName, ItemName, MoveName } from '@smogon/calc/dist/data/int
 import type { CalcdexBattleState, CalcdexPokemon, CalcdexPokemonPreset } from '@showdex/redux/store';
 import { detectGenFromFormat } from './detectGenFromFormat';
 import { detectLegacyGen } from './detectLegacyGen';
-import { getDexForFormat } from './getDexForFormat';
 import { mergeRevealedMoves } from './mergeRevealedMoves';
 import { sanitizePokemon } from './sanitizePokemon';
 import { sanitizeMoveTrack } from './sanitizeMoveTrack';
