+++ src/utils/calc/calcMoveBasePower.ts
@@ -1,11 +1,10 @@
 import { formatId } from '@showdex/utils/app';
-import { getDexForFormat } from '@showdex/utils/battle';
 import { clamp } from '@showdex/utils/core';
+import { getDexForFormat, shouldBoostTeraStab } from '@showdex/utils/dex';
 import type { MoveName } from '@smogon/calc/dist/data/interface';
 import type { CalcdexPokemon } from '@showdex/redux/store';
 import type { SmogonMoveOverrides } from './createSmogonMove';
 import { calcHiddenPower } from './calcHiddenPower';
-import { shouldBoostTeraStab } from './shouldBoostTeraStab';
 
 /**
  * Calculates the base power of the provided `moveName` based on conditions of the `pokemon`.
