+++ src/utils/calc/index.ts
@@ -1,4 +1,3 @@
-export * from './alwaysCriticalHits';
 export * from './calcCalcdexId';
 export * from './calcCalcdexNonce';
 export * from './calcHiddenPower';
@@ -12,14 +11,8 @@ export * from './convertLegacyStats';
 export * from './createSmogonField';
 export * from './createSmogonMove';
 export * from './createSmogonPokemon';
-export * from './determineCriticalHit';
-export * from './determineMoveTargets';
 export * from './findHighestStat';
-export * from './getMoveOverrideDefaults';
 export * from './guessServerLegacySpread';
 export * from './guessServerSpread';
-export * from './hasMoveOverrides';
-export * from './shouldBoostTeraStab';
-export * from './shouldIgnoreItem';
 export * from './statModRecorder';
 export * from './useSmogonMatchup';
