+++ src/utils/dex/alwaysCriticalHits.ts
@@ -1,7 +1,7 @@
 import { PokemonCriticalHitMoves } from '@showdex/consts/pokemon';
-import { getDexForFormat } from '@showdex/utils/battle';
 import type { GenerationNum } from '@smogon/calc';
 import type { MoveName } from '@smogon/calc/dist/data/interface';
+import { getDexForFormat } from './getDexForFormat';
 
 /**
  * Internally used regex for testing if the move's description (either `desc` or `shortDesc`) mentions
