+++ src/utils/dex/formatDexDescription.ts
