+++ src/utils/dex/getZMove.ts
@@ -3,7 +3,7 @@ import { formatId } from '@showdex/utils/app';
 import { logger } from '@showdex/utils/debug';
 import type { ItemName, MoveName } from '@smogon/calc/dist/data/interface';
 
-const l = logger('@showdex/utils/app/getZMove');
+const l = logger('@showdex/utils/dex/getZMove');
 
 /**
  * Returns the corresponding Z move for a given move.
