+++ src/utils/dex/hasNickname.ts
